Extract the three subscribe_ methods to separate module
